[add] exception
