--- src/main/java/com/duroop/dodom/appointment/entity/Appointment.java
@@ -1,33 +1,0 @@
-package com.duroop.dodom.appointment.entity;
-
-import com.duroop.dodom.audit.BaseTimeEntity;
-import com.duroop.dodom.counselor.entity.Counselor;
-import com.duroop.dodom.review.entity.Review;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.*;
-import org.springframework.format.annotation.DateTimeFormat;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter @Setter
-@Builder
-public class Appointment extends BaseTimeEntity {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long appointmentId;
-    private String email;
-    @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
-    private LocalDateTime date;
-    private String inquiry;
-    private String method;
-
-    @OneToOne(mappedBy = "appointment", orphanRemoval = true)
-    private Review review;
-
-    @ManyToOne @JoinColumn(name = "counselor_id")
-    private Counselor counselor;
-}
