--- src/main/java/com/duroop/dodom/appointment/repository/AppointmentRepository.java
@@ -1,10 +1,0 @@
-package com.duroop.dodom.appointment.repository;
-
-import com.duroop.dodom.appointment.entity.Appointment;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
-
-}
