+++ src/main/java/com/duroop/dodom/domain/appointment/repository/AppointmentRepository.java
@@ -1,0 +1,10 @@
+package com.duroop.dodom.domain.appointment.repository;
+
+import com.duroop.dodom.domain.appointment.entity.Appointment;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
+
+}
