+++ src/main/java/com/duroop/dodom/domain/review/entity/Review.java
@@ -1,0 +1,29 @@
+package com.duroop.dodom.domain.review.entity;
+
+import com.duroop.dodom.domain.appointment.entity.Appointment;
+import com.duroop.dodom.audit.BaseTimeEntity;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.*;
+
+@Entity
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter @Setter
+public class Review extends BaseTimeEntity {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long reviewId;
+    private String name;
+    private double score;
+    private String review;
+    private String review1;
+    private Boolean agree = true;
+
+    @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    @JoinColumn(name = "appointment_id")
+    private Appointment appointment;
+}
