+++ src/main/java/com/duroop/dodom/exception/BusinessException.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class BusinessException extends RuntimeException{
 
-    private ExceptionCode exceptionCode;
+    private final ExceptionCode exceptionCode;
 
     public BusinessException(ExceptionCode exceptionCode) {
         super(exceptionCode.getMessage());
