+++ src/main/java/com/duroop/dodom/exception/ExceptionResponse.java
@@ -1,0 +1,24 @@
+package com.duroop.dodom.exception;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@Getter @Setter
+public class ExceptionResponse {
+
+    private int code;
+    private String message;
+
+    public ExceptionResponse(ExceptionCode exceptionCode) {
+        this.code = exceptionCode.getCode();
+        this.message = exceptionCode.getMessage();
+    }
+
+    public static ExceptionResponse of(ExceptionCode exceptionCode) {
+        return new ExceptionResponse(exceptionCode);
+    }
+}
