--- src/main/java/com/duroop/dodom/score/repository/ScoreRepository.java
@@ -1,8 +1,0 @@
-package com.duroop.dodom.score.repository;
-
-
-import com.duroop.dodom.score.entity.Score;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ScoreRepository extends JpaRepository<Score, Long> {
-}
