--- src/main/java/com/duroop/dodom/tag/repository/TagRepository.java
@@ -1,8 +1,0 @@
-package com.duroop.dodom.tag.repository;
-
-import com.duroop.dodom.tag.entity.Tag;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface TagRepository extends JpaRepository<Tag, Long> {
-}
